Remap addresses so that the recompiler cache doesn't have to worry about banking or mirrors

+++ core/DMGRecompiler.h
@@ -34,8 +34,8 @@ private:
     {
         CompiledFunc func;
         uint8_t *endPtr;
-        uint16_t endPC;
+        uint32_t endPC;
     };
 
-    std::map<uint16_t, FuncInfo> compiled;
+    std::map<uint32_t, FuncInfo> compiled;
 };
